renamed product validation file
