+++ customers.db
@@ -1,1 +1,2 @@
+{"firstName":"Test","lastName":"Tester","email":"testman@testmail.com","password":"thisisatest","phoneNumber":"101010010011","loggedIn":false,"_id":"5ZMAwqKTd5acajBT"}
 {"firstName":"Guest","loggedIn":true,"_id":"guestintest"}
