+++ orderHistory.db
@@ -1,0 +1,1 @@
+{"userId":"5ZMAwqKTd5acajBT","firstName":"Test","totalPrice":105.3,"orders":[{"orderId":290631,"userId":"5ZMAwqKTd5acajBT","items":[{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"_id":"nZDt5gDMOZGl4nCC"},{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"_id":"nZDt5gDMOZGl4nCC"},{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"_id":"nZDt5gDMOZGl4nCC"}],"totalPrice":105.3,"orderTime":"2024-06-06 18:48:14","deliveryTime":"2024-06-06 19:08:14"}],"_id":"D5k7OfMykrc2VrHa"}
